fix(croppingPlanes): use signs of direction vectors
When the direction vectors of an image "mirror" the spacing,
the image cropping planes were outside the image.
